@@ -13,7 +13,6 @@
     "chai-arrays": "^2.0.0",
     "chai-http": "^4.2.0",
     "dotenv": "^6.2.0",
-    "eslint": "^5.11.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
@@ -23,7 +22,6 @@
     "joi": "^14.3.1",
     "jsonwebtoken": "^8.4.0",
     "make-runnable": "^1.3.6",
-    "mocha": "^5.2.0",
     "moment": "^2.23.0",
     "nodemon": "^1.18.6",
     "pg": "^7.8.0",
@@ -56,8 +54,11 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-stage-2": "^7.0.0",
+    "@types/express": "^4.16.1",
     "babel-preset-env": "^1.7.0",
     "babel-register": "^6.26.0",
+    "eslint": "^5.13.0",
+    "mocha": "^5.2.0",
     "nyc": "^13.1.0",
     "webpack": "^4.29.3"
   }
