@@ -5,19 +5,19 @@ import dotenv from 'dotenv';
 dotenv.config();
 
 const pool = new Pool({
-	connectionString: process.env.DATABASE_URL
+  connectionString: process.env.DATABASE_URL
 });
 
 module.exports = {
-	query(text, params){
-		return new Promise((resolve, reject) => {
-			pool.query(text, params)
-				.then((res) => {
-					resolve(res);
-				})
-				.catch((err) => {
-					reject(err);
-				});
-		});
-	}
+  query(text, params){
+    return new Promise((resolve, reject) => {
+      pool.query(text, params)
+        .then((res) => {
+          resolve(res);
+        })
+        .catch((err) => {
+          reject(err);
+        });
+    });
+  }
 };
