@@ -5,41 +5,41 @@ import jwt from 'jsonwebtoken';
 import key from '../helpers/keys';
 
 const Helper = {
-	/**
+  /**
    * Hash Password Method
    * @param {string} password
    * @returns {string} returns hashed password
    */
-	hashPassword(password) {
-		return bcrypt.hashSync(password, bcrypt.genSaltSync(8));
-	},
-	/**
+  hashPassword(password) {
+    return bcrypt.hashSync(password, bcrypt.genSaltSync(8));
+  },
+  /**
    * comparePassword
    * @param {string} hashPassword 
    * @param {string} password 
    * @returns {Boolean} return True or False
    */
-	comparePassword(hashPassword, password) {
-		return bcrypt.compareSync(password, hashPassword);
-	},
-	/**
+  comparePassword(hashPassword, password) {
+    return bcrypt.compareSync(password, hashPassword);
+  },
+  /**
    * isValidEmail helper method
    * @param {string} email
    * @returns {Boolean} True or False
    */
-	isValidEmail(email) {
-		return /\S+@\S+\.\S+/.test(email);
-	},
-	/**
+  isValidEmail(email) {
+    return /\S+@\S+\.\S+/.test(email);
+  },
+  /**
    * Gnerate Token
    * @param {string} id
    * @returns {string} token
    */
-	generateToken(payload) {
-		const token = jwt.sign(payload, key.secret
-		);
-		return token;
-	}
+  generateToken(payload) {
+    const token = jwt.sign(payload, key.secret
+    );
+    return token;
+  }
 };
 
 export default Helper;
