@@ -1,3 +1,4 @@
-const Keys={    secret:'aaron'
+const Keys={
+	secret:'aaron'
 };
 export default Keys;
